refactor: Move PDA derivation tests to separate file
Extracted all PDA derivation related tests from `utils.ts` into a new `pdas.ts` test file for better organization and separation of concerns.
